feat: 메일 보관함 기능

+++ backend/src/main/java/com/example/backend/mail/model/repository/EmailRepository.java
@@ -3,5 +3,7 @@ package com.example.backend.mail.model.repository;
 import com.example.backend.mail.model.entity.Email;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailRepository extends JpaRepository<Email, Integer> {
+import java.util.List;
+
+public interface EmailRepository extends JpaRepository<Email, Long> {
 }
